Wrap code in a define block only if export config exists
Modules from NPM expect `module` and `exports` to exist.
The current temp hack to not wrap code in a define block
if a package.json exists actually stops true node supporting
modules working altogether.

I've added a "excludeDefine" library.manifest config option. This is
required for browser-modules because:

* it is in npm and has a package.json
* it declares the `define` functions so can't be wrapped in `define`

This means that browser-modules can be included and not be
wrapped, but anything else that defines a package.json is wrapped
and handles its own exporting; it's a node module from npm,
that's how they work.

I've also added something so that `define` and  `module.exports`
are only written if something has been defined to be exported via
the `exports` config. This was previously a @NotNull and @NotEmpty
config value. But since browser-modules has no use for an `exports`
this seemed like there were occasions where nothing will be exported
via `module.exports`.