Add test version of XMLResourceService
It was missing which caused our tests to fail as they couldn't load the
XML resources.

Note that it uses the old URL (xml.bundle instead of bundle.xml). This
is after discussing with @sospirited ... he said you guys will change it
to the correct URL once he verifies that BRJS is backwardly compatibale
in this regard.