removing the overriddden JAVA_OPTS from brjs
removing the overdden JAVA_OPTS from the brjs commands since nowadays
the default JAVA_OPTS are sufficient and people generally have their own
custom settings of it anyway. This both fixes the problem of JAVA_OPTS
getting longer and the MaxPermSize warning caused in Java8.

see
https://stackoverflow.com/questions/16095439/out-of-memory-error-java-heap-size-when-memory-is-available
for info on the default values. fixes #518