Started updating all classes to:
  * Use a @module annotation, since all classes are now exported as
modules
  * Use an @alias annotation if the module merely contains a single
class.
  * Use a @classdesc annotation to explicitly delimit a classes textual
description, since not doing so leads to inconsistencies depending on
whether a @module annotation is being used or not, and having to have
the class description text ahead of the annotations that describe the
class make it harder to read.
  * Refer to all other classes as <module>:<require-path>.
  * Place the constructor annotations at the very bottom, as close to
the method arguments as possible.

This work is being paused since I've noticed that some of the classes
I'm trying to update have since been deleted, so it would be worth
merging 'develop' before continuing.