Squashed 'brjs-sdk/build-resources/includes/sdk/jsdoc-toolkit-resources/jsdoc-toolkit/' changes from 06ad31c..09d199b
09d199b Work-around for <https://github.com/jsdoc3/jsdoc/issues/763> that allows our NamespacedJs style classes to have their methods and static constants to continue to be displayed without breaking any of the existing spec tests -- static methods will not currently display.
f8d61e9 Merging changes from 'index-filtering-support' branch
765b3ec fixed remaining linting errors.
01530b2 fix all linting errors except 'no-use-before-define', since these might cause merge conflicts if mixed in with the other error fixes.
ef21fed Work around the bug that occurs in IE10 and below (pressing enter in a text box causes the first button in the same form to be clicked) by explicitly wrapping in a <form/> tag in case other buttons are added to the page (we do this in the BladeRunnerJS template), and ensuring that the other button in the same form is not considered to be a submit button.
b236436 Clean up how we are doing the browser specific fix-ups so they are only applied on the relevant browsers.
ea3dd02 Prevent IE8 from trying to submit when the enter key is pressed.
f9d3ed5 Use a PNG instead of an SVG for the 'x' button so that it can be viewed in IE8.
923d4e8 Merge commit 'ba93ca08813d89e79927bf514208c8a994c7984d' into jsdoc-index-filtering-support
898e2b9 Merge branch 'jsdoc-improvements' into jsdoc-index-filtering-support
cda906e re-added html5shiv as the style stopped working on IE8 after I removed it.
c7c968d Merge branch 'rebase-typography-off-of-mdn' into brjs
312d4e5 Don't allow the title to overflow when using a lower resolution display.

git-subtree-dir: brjs-sdk/build-resources/includes/sdk/jsdoc-toolkit-resources/jsdoc-toolkit
git-subtree-split: 09d199b7bb52571f392821be970028f980a6725f