Stopped injecting the FileModificationService as a dependency into BRJS, and instead assumed the OptimisticFileModificationService by default, but where BRJS.setFileModificationService() can be invoked if this is not suitable. As a consequence of us know using the OptimisticFileModificationService for both test and non-test environments, created a new TimeAccessor dependency that must be injected into BRJS, and will be different in test and not-test environments.
The BRJS.setFileModificationService() is complicated by the fact that
many FileInfo objects will already have been issued by the time
etFileModificationService() is invoked, and since these FileInfo objects
wrap a FileModificationInfo object that will have been created by the
previous FileModificationService, we need a way to dynamically swap the
implementations around. This is done using the BRJSFileInfo class, which
proxies through to an underlying FileInfo object that can be swapped out
as necesarry.