First version of the inline plugin.
In some circumstances it is useful to copy the contents of a file into another file. This can be especially useful when trying to reduce the perceived load time of a page by optimising the critical render path. Critical JS and CSS can now easily be inlined into the index.html page using this plugin.

Example of usage:

<@inline file=“../libs/mylibrary/src/mylibrary/FileToInline.js” @/>

The ‘file’ argument will be relative to the file that includes the ‘inline’ tag.