We need a selectionField which does not pick up "random" value when options are changed. Instead we need to force user to repick value and throw validation error.
Use case:
Imagine you have accounts loaded in your app. You start a trade with a list of accounts from the model. But while you are in a middle of the trade your permission for selected account is rejected. So your permissioned accounts changed. Current behaviour will pick first account in list and populate your model with it. This will result in you trading on different account then you originally intended (which can be big mistake). This patch will set value of account in a model to "" which will lead into validation error and your trade will get rejected.

This is also fixing a bug where code bellow is effectively comparing labels of options with value of selectionField, which doesn't work when label is different then value of option:
var mOptions = br.util.MapUtility.addArrayToMap({}, pOptions);
if(!mOptions[this.value.getValue()])