fixing a bug where files with £ werent renamespaced
this was caused due to MimeMagic throwing an exception and/or not seeing
the mime type as a text file. The root cause of the issue the use of
MimeMagic was trying to fix was a String == String comparrision when
trying to detirmine whether the content has changed. This has now been
fixed so we dont need to use MimeMagic anymore