Temporarily disabled part of BRAppMetaService
It was sneakilly trying to require a dependency that it had hidden from
BRJS on the server. This is needed so that I can progress the spike, and
since with SystemJS it will no longer be able to hide a dependency on
the server, but then require it on the client.