Don't consider `__` to be private
Stopped considering double underscore as meaning private, since this is
normally used by JavaScript developers as a way of more safely adding
properties to other peoples objects in a way that's far less likely to
cause name collision.

Specifically, this was causing Knockout to fail since it uses the
presence of a `__ko__` property to determine whether an an argument is
a Knockout observable.