This patch adds a flag to track whether validation is currently in progress. This is important for fields where the validation is asynchronous. We generally do not want to be displaying an error while this is occurring, but we also want to ensure we have a mechanism to disable other functionality until the field has been validated.
Also included in this patch is a hasFocus attribute, which can be used by controls to indicate that the field is focused. This is useful for date fields, as the field is not validated until it has lost focus. When this is the case we want to disable other functionality until we know that the date is valid.