issue #1632: ensures all depended resources are 	bundled when accessing workbenches.
The fix for #1517 appears valid (NPE's no longer appear to be an issue).

However, some resources were not bundled due to a primary require path collision.

Assets at blade level have the same primary require path as those in a
blade's workbench.

The primary require path was being used as a guard to
prevent multiple instances of the same asset being added to the
bundleset (BundleSetBuilder#addLinkedAsset).

Changing the guard to use the full asset path of the asset, prevents
this collisions and allows all assets to be processed.

Spec tests have been added to validate the fix.