Add ability to require index.js files from the root asset container.
The assetRegistry contains a number of synthetic assets, which are used
as a way to add implicit dependencies for other assets, but are not
useful as requirable assets themselves.

We wish to allow the requiring of index.js assets from these locations,
but the existing require paths prevented their re-use.

This change (along with the previous commit) allows the
differentiation between these synthetic assets (DirectoryAsset and
BRJSConformantRootDirectoryAsset) and an index file contained within
them.

DirectoryAsset require paths are suffixed with: @dir
BRJSConfirmantRootDirectoryAsset require paths are suffixed with: @root

A spec test suite has been added to validate new require functionality.