refactor(jquery): Remove module export block
Removing block that exports jQuery when it detects that `module` exists.
In webpack this evaluated to true and caused `jQuery` not to be
exported to the window. The jquery UI scripts that are included in the
jQuery thirdparty library bundle depend on `jQuery` being availble on
the global. Setting `module` to null meant that webpack couldn't
then export `jQuery` from the converted library module.

closes: CTNXT-251